Added settings.html; will need further editing
While I made the div a <form> and added buttons and labels, the settings.html page would require maybe JavaScript or something else in order for the Settings aspects to be truly functional.  Like the other pages, it will require styling that would match most settings aspects in other websites; it will be simple, but also have enough elements and features for it to work.  As of now, I am looking at more examples of settings and how to implement them to HTML and CSS.  Some of the labels and buttons can be renamed so they each have their own function -- maybe a lever code would be added.
If permissible, I can draw out a blueprint on what the Settings page would look like, and then have it implemented into HTML.