Added signup.html page; requires further editing
My logic is that if a website has a Login page, a Sign Up page may be required as well.  I reused the code from login.html and added extra elements (Email, Confirm Password, etc.) along with a comment suggesting extra options for logging in.  The checkmark can be an "I Understand the Terms of this Site" or "I am not a robot;" or maybe this page doesn't need a checkmark?
Like all the other HTML pages, the signup.html page will require styling and reformatting not only to make it look nice, but also help with its overall functionality.

Unless there is anything else I need to add and/or create/prepare, I will not be adding any more webpages for the website, and will stick to proposing styles, visuals and/or text/proofreading aspects as well as more blueprinting.  Some time around April (1st week or 2nd week) hopefully we will have everything done before submitting the project.